Cleanup Compressor.h
